[ISSUE-164] refactor durations
